moved qt.py into shared folder
