+++ thoughtful_termites/app/__init__.py
@@ -3,7 +3,7 @@ def run():
     Convenience function for running the app.
     """
     import sys
-    from thoughtful_termites.app import qt
+    from thoughtful_termites.shared import qt
     from thoughtful_termites.app.widgets import GoalListWindow
 
     app = qt.QApplication(sys.argv)
