+++ thoughtful_termites/app/widgets/edit_daily_reminder_window.py
@@ -1,4 +1,4 @@
-from .. import qt
+from thoughtful_termites.shared import qt
 
 
 class EditDailyReminderWindow(qt.QDialog):
