+++ thoughtful_termites/app/widgets/edit_goal_window.py
@@ -1,4 +1,4 @@
-from .. import qt
+from thoughtful_termites.shared import qt
 from ..goal_suggestions import get_suggestion
