+++ thoughtful_termites/shared/qt.py
